Make cassandra integration compatible with both guava 20 and 27
The issue seems to be related to invocation of guava's `Futures.tranform()` which existed in 2 forms in guava 20.x
and only in one form in guava 27.

Specifically:
[Guava 20](https://google.github.io/guava/releases/20.0/api/docs/index.html?help-doc.html):
- `transform(ListenableFuture<I> input, Function<? super I,? extends O> function)` <<-- used by our code as of now.
- `transform(ListenableFuture<I> input, Function<? super I,? extends O> function, Executor executor)`
While [Guava 27]():
- Only `transform(ListenableFuture<I> input, Function<? super I,? extends O> function, Executor executor)`

A possible approach to be tested would be to retrieve an `Executor` and pass it to the new method.
The way the executor was retrieved in Guava 20
[is quite simple](https://github.com/google/guava/blob/65f6b4f4b132a051616403bcf74e4034a19d92a1/guava/src/com/google/common/util/concurrent/MoreExecutors.java#L425)
and this hasn't actually changed until guava 27.